+++ src/life.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <webgpu/webgpu.h>
-#include <vector>
 #include <memory>
 #include "WebGPUContext.h"
 #include "Geometry.h"
@@ -10,7 +9,6 @@
 class Life
 {
 public:
-    static constexpr uint32_t GRID_SIZE = 32;
 
     class InitializationError : public std::runtime_error {
         public:
@@ -31,42 +29,5 @@ private:
 
     // Component Classes
     std::unique_ptr<Geometry> geometry; 
-    std::unique_ptr<RenderPipeline> pipeline; 
-    
-    // Uniform Buffer & Bindgroup
-    WGPUBuffer uniformBuffer = nullptr;
-    WGPUBindGroup uniformBindGroup = nullptr;
-
-    // Cell State Storage Buffer
-    WGPUBuffer cellStateStorageBufferA = nullptr;
-    WGPUBuffer cellStateStorageBufferB = nullptr;
-    std::vector<uint32_t> cellStateArray;
-    
-    // Shader module and render pipeline
-    WGPURenderPipeline cellPipeline = nullptr;
-
-    // Bind group
-    WGPUBindGroupLayout bindGroupLayout = nullptr;
-    WGPUBindGroup bindGroups[2] = {nullptr, nullptr};
-    
-    // Surface state
-    bool surfaceCreated = false;
-    int width = 800;   // Canvas dimensions
-    int height = 600;
-    
-    static constexpr uint32_t INSTANCE_COUNT = GRID_SIZE * GRID_SIZE;
-    static constexpr float UNIFORM_ARRAY[2] = {
-        static_cast<float>(GRID_SIZE), 
-        static_cast<float>(GRID_SIZE)
-    };
-
-    // Frame count
-    uint32_t frameCount = 0;
-
-    // Internal methods
-    void createUniformBuffer();
-    void createRenderPipeline();
-    void createCellStateStorageBuffer();
-    void createBindGroupLayout();
-    void createBindGroup();
+    std::unique_ptr<RenderPipeline> pipeline;    
 };
